shadow mapping

+++ src/dxerr.hpp
@@ -5,6 +5,6 @@
 #define DXERR(x, msg)                                                                              \
     if (FAILED(x))                                                                                 \
     {                                                                                              \
-        spdlog::error("{}: 0x{:x}", msg, x);                                                       \
+        spdlog::error("{}: 0x{:x}", msg, static_cast<unsigned long>(x));                           \
         return false;                                                                              \
     }
